+++ apps/server/package.json
@@ -3,7 +3,6 @@
   "type": "module",
   "version": "1.0.0-beta.21",
   "exports": {
-    ".": "./src/index.ts",
     "./types": "./src/types.ts"
   },
   "scripts": {
