+++ apps/server/src/ws/index.ts
@@ -1,0 +1,2 @@
+export type * from './events'
+export type * from './routes'
